+++ pkgs/ffigen/lib/src/code_generator/typealias.dart
@@ -157,6 +157,10 @@ class Typealias extends BindingType {
   }
 
   @override
+  String getObjCBlockSignatureType(Writer w) =>
+      type.getObjCBlockSignatureType(w);
+
+  @override
   bool get sameFfiDartAndCType => type.sameFfiDartAndCType;
 
   @override
@@ -170,8 +174,14 @@ class Typealias extends BindingType {
     Writer w,
     String value, {
     required bool objCRetain,
+    required bool objCAutorelease,
   }) =>
-      type.convertDartTypeToFfiDartType(w, value, objCRetain: objCRetain);
+      type.convertDartTypeToFfiDartType(
+        w,
+        value,
+        objCRetain: objCRetain,
+        objCAutorelease: objCAutorelease,
+      );
 
   @override
   String convertFfiDartTypeToDartType(
@@ -231,8 +241,9 @@ class ObjCInstanceType extends Typealias {
     Writer w,
     String value, {
     required bool objCRetain,
+    required bool objCAutorelease,
   }) =>
-      ObjCInterface.generateGetId(value, objCRetain);
+      ObjCInterface.generateGetId(value, objCRetain, objCAutorelease);
 
   @override
   String convertFfiDartTypeToDartType(