+++ pkgs/objective_c/CHANGELOG.md
@@ -5,6 +5,8 @@
 - Added `ObjCBlock`, which is the new user-facing representation of ObjC blocks.
 - Migrate to ARC (Automatic Reference Counting).
 - Enable ObjC objects and blocks to be sent between isolates.
+- Add `autorelease` and `retainAndAutorelease` methods to ObjC objects and
+  blocks.
 - __Breaking change__: Remove some convenience methods from `_ObjCRefHolder`:
   `isReleased`, `release`, `pointer`, and `retainAndReturnPointer`. Uses of
   these methods now need to go through `.ref`. Eg `obj.pointer` becomes